Move species compendia to `/pages` dir
