+++ src/pages/Main/EmailSection.js
