+++ src/pages/SpeciesCompendia/PointsBackground.js
